+++ Sources/Bl00p/Models.swift
@@ -431,6 +431,8 @@ struct ManagerWorkflow: Identifiable, Codable, Hashable, Sendable {
     var managerProfileID: UUID
     var team: ManagerTeamConfiguration
     var request: String
+    var implementationPlan: String?
+    var planApprovalEntryID: UUID?
     var stage: ManagerWorkflowStage
     var branch: String?
     var pullRequestURL: String?
@@ -445,6 +447,8 @@ struct ManagerWorkflow: Identifiable, Codable, Hashable, Sendable {
         managerProfileID: UUID,
         team: ManagerTeamConfiguration,
         request: String,
+        implementationPlan: String? = nil,
+        planApprovalEntryID: UUID? = nil,
         stage: ManagerWorkflowStage = .planning,
         branch: String? = nil,
         pullRequestURL: String? = nil,
@@ -458,6 +462,8 @@ struct ManagerWorkflow: Identifiable, Codable, Hashable, Sendable {
         self.managerProfileID = managerProfileID
         self.team = team
         self.request = request
+        self.implementationPlan = implementationPlan
+        self.planApprovalEntryID = planApprovalEntryID
         self.stage = stage
         self.branch = branch
         self.pullRequestURL = pullRequestURL